refactor: MaybeEntryをShared空間に移動
MaybeEntryをParseコマンドだけでなくStringifyコマンドでも使う予定のため。
